7a: split five test suites across the tier boundary
Five suites turned out to be mixed on closer read (invariant 8's warning about
trusting a using-directive census applies to tests too) - each pure half moved
under the naming rule from e677b54: the half with more test methods keeps the
original name, the smaller half takes a new one.

RecognizerExecutorTests (Tests.Core, kept name, 23 methods) vs the
GrabTemplateExecutor-backed recognizer-placeholder tests (7 methods) - merged
into the existing Tests/GrabTemplateExecutorTests.cs instead of a new class,
since GrabTemplateExecutor needs System.Windows.Rect and that file already
covers it comprehensively.

PatternExecutorTests (Tests.Core, kept name, 10 methods) vs
PatternItemCatalogTests (Tests, new, 3 methods) - PatternItemCatalog is the
app-side half of the PatternItem split from e677b54.

ProtocolUtilitiesTests (Tests.Core, kept name, 9 methods) vs
ProtocolHandlerUtilitiesTests (Tests, new, 6 methods) - ProtocolHandlerUtilities
is internal and app-side (path validation against the filesystem).

BarcodeUtilitiesTests (Tests.Core.Windows, kept name, 3 methods) vs
BarcodeUtilitiesImageTests (Tests, new, 1 method) - the one method is
[WpfFact]-tagged; Xunit.StaFact cannot be referenced from Tests.Core.Windows,
confirmed empirically by loading its net8.0-windows7.0 assembly and checking
GetReferencedAssemblies() - it names WindowsBase and System.Windows.Forms,
which TierBoundaryTests bans.

FilesIoTests (Tests, kept name, 9 methods stay app/WPF-bound) split two ways:
IoUtilitiesTests (Tests.Core, new, 3 methods against plain-Core IoUtilities)
and FileUtilitiesTests (Tests.Core.Windows, new, 1 method against
FileUtilities.GetVisualDocumentFilter).

OcrTests is the big one (27 methods vs 15). The pure half kept the OcrTests
name and moved to Tests.Core.Windows; the app-coupled half - OcrSourceUtilities
calls, WPF BitmapImage construction, direct AppUtilities.TextGrabSettings reads
- became Tests/OcrSourceTests.cs. Four methods (OcrComplexTableTestImage,
GetTessLanguages, GetTesseractStrongLanguages, GetTesseractGitHubLanguage) were
tagged [WpfFact] in the original but never touch a WPF type once OcrUtilities/
TesseractHelper/TesseractGitHubFileDownloader had already moved to
Core/Core.Windows in earlier waves; they moved as [Fact]/[Fact(Skip=...)] with
no behavior change; [Fact(Skip=...)] tests still compile-check but skip their
body, so this only mattered for OcrComplexTableTestImage, which now runs.

Two more things came out of reading OcrUtilities in full rather than trusting
the method-body census: BuildTextFromOcrLines reads SettingsAccess.Current
unconditionally as the first operand of a short-circuited && (so the four
BuildTextFromOcrLines_* tests need a resolver even though their assertions
don't obviously depend on a setting value), and this project's own namespace
ending in ".Windows" makes an unqualified "Windows.Foundation.Rect" resolve
against itself and fail to compile - fixed with an explicit
"using Windows.Foundation;" and dropping the qualification.

Tests.Core.Windows/FakeTextGrabSettings.cs is the resolver those four tests
need: a plain ITextGrabSettings POCO seeded from Settings.settings's shipped
defaults (RemoveFurigana=true matters concretely - one test's expected output
depends on furigana actually being filtered), registered via
[ModuleInitializer] the same way the app's SettingsAccessInitializer does.

Table-Complex-WordBorders.json moved with OcrComplexTableTestImage since
nothing else reads it; Table-Complex.png stayed put (and unused - dead even
before this batch, per its own unreferenced private const).
