+++ Tests.Core.Windows/TextFiles/Table-Complex-WordBorders.json
