+++ Tests.Core/PatternExecutorTests.cs
@@ -1,8 +1,13 @@
 using Text_Grab.Models;
 using Text_Grab.Utilities;
 
-namespace Tests;
+namespace Text_Grab.Tests.Core;
 
+// Pure half of the original Tests/PatternExecutorTests.cs (batch 7a): PatternExecutor and
+// StoredRegex-backed PatternItem construction are Core-only. The three PatternItemCatalog
+// tests stayed behind as Tests/PatternItemCatalogTests.cs - PatternItemCatalog.GetAll()/
+// GetByName() are app-side (Text-Grab/Models/PatternItemCatalog.cs, per e677b54). This half
+// has 10 methods against that one's 3, so it kept the original name.
 public class PatternExecutorTests
 {
     // A deterministic saved-regex item that does not depend on the machine's saved patterns.
@@ -11,45 +16,6 @@ public class PatternExecutorTests
 
     private static PatternItem RecognizerByName(string name) =>
         new(BuiltInRecognizer.GetByName(name) ?? throw new InvalidOperationException($"missing recognizer {name}"));
-
-    // ── PatternItem catalog ───────────────────────────────────────────────────
-
-    [Fact]
-    public void GetAll_ListsSavedRegexesBeforeRecognizers()
-    {
-        IReadOnlyList<PatternItem> all = PatternItemCatalog.GetAll();
-
-        int firstRecognizer = -1;
-        int lastSaved = -1;
-        for (int i = 0; i < all.Count; i++)
-        {
-            if (all[i].Kind == PatternKind.Recognizer && firstRecognizer < 0)
-                firstRecognizer = i;
-            if (all[i].Kind == PatternKind.SavedRegex)
-                lastSaved = i;
-        }
-
-        Assert.True(firstRecognizer >= 0, "expected at least one recognizer item");
-        Assert.True(lastSaved < firstRecognizer, "all saved regexes should precede recognizers");
-    }
-
-    [Fact]
-    public void GetAll_IncludesEveryRecognizerWithSmartGroup()
-    {
-        List<PatternItem> recognizers = [.. PatternItemCatalog.GetAll().Where(p => p.Kind == PatternKind.Recognizer)];
-
-        Assert.Equal(BuiltInRecognizer.GetAll().Count, recognizers.Count);
-        Assert.All(recognizers, p => Assert.Equal(PatternItem.SmartGroup, p.GroupLabel));
-    }
-
-    [Fact]
-    public void GetByName_FindsRecognizer_CaseInsensitive()
-    {
-        PatternItem? email = PatternItemCatalog.GetByName("EMAIL");
-
-        Assert.NotNull(email);
-        Assert.Equal(PatternKind.Recognizer, email!.Kind);
-    }
 
     // ── PatternExecutor – recognizer-backed ───────────────────────────────────
