+++ Tests/FilesIoTests.cs
@@ -7,6 +7,11 @@ using Text_Grab.Utilities;
 
 namespace Tests;
 
+// App-coupled remainder of the original file (batch 7a). Its IoUtilities-only tests moved to
+// Tests.Core/IoUtilitiesTests.cs and its FileUtilities.GetVisualDocumentFilter test moved to
+// Tests.Core.Windows/FileUtilitiesTests.cs. What is left here is blocked on WPF ([WpfFact]/
+// [WpfTheory] needing Xunit.StaFact, which cannot be referenced outside Tests) or on the
+// app-side OpenDocumentFilterUtilities/App classes, so it kept the original FilesIoTests name.
 public class FilesIoTests
 {
     private const string fontSamplePath = @"Images\font_sample.png";
@@ -95,48 +100,6 @@ public class FilesIoTests
         string fileName = "FileNotFound.json";
         Bitmap? emptyReturn = await FileUtilities.GetImageFileAsync(fileName, storageKind);
         Assert.Null(emptyReturn);
-    }
-
-    [Theory]
-    [InlineData(@"C:\Temp\sheet.csv", EtwEditorMode.Spreadsheet)]
-    [InlineData(@"C:\Temp\sheet.TSV", EtwEditorMode.Spreadsheet)]
-    [InlineData(@"C:\Temp\sheet.tab", EtwEditorMode.Spreadsheet)]
-    [InlineData(@"C:\Temp\notes.md", EtwEditorMode.Markdown)]
-    [InlineData(@"C:\Temp\notes.markdown", EtwEditorMode.Markdown)]
-    [InlineData(@"C:\Temp\notes.txt", EtwEditorMode.Text)]
-    [InlineData(@"C:\Temp\data.json", EtwEditorMode.Text)]
-    public void GetEditorModeForPath_UsesFileExtension(string path, EtwEditorMode expectedMode)
-    {
-        Assert.Equal(expectedMode, IoUtilities.GetEditorModeForPath(path));
-    }
-
-    [Theory]
-    [InlineData(@"C:\Temp\scan.png", OpenContentKind.Image)]
-    [InlineData(@"C:\Temp\scan.PDF", OpenContentKind.PdfDocument)]
-    [InlineData(@"C:\Temp\notes.txt", OpenContentKind.TextFile)]
-    public void GetOpenContentKindForPath_ClassifiesVisualDocumentsAndText(string path, OpenContentKind expectedKind)
-    {
-        Assert.Equal(expectedKind, IoUtilities.GetOpenContentKindForPath(path));
-    }
-
-    [Theory]
-    [InlineData(".png", true)]
-    [InlineData(".PDF", true)]
-    [InlineData(".txt", false)]
-    [InlineData("", false)]
-    public void IsVisualDocumentFileExtension_RecognizesImagesAndPdf(string extension, bool expected)
-    {
-        Assert.Equal(expected, IoUtilities.IsVisualDocumentFileExtension(extension));
-    }
-
-    [Fact]
-    public void GetVisualDocumentFilter_IncludesPdfSupport()
-    {
-        string filter = FileUtilities.GetVisualDocumentFilter();
-
-        Assert.Contains("Image and PDF files|", filter);
-        Assert.Contains("PDF files|*.pdf", filter);
-        Assert.Contains("Image files|", filter);
     }
 
     [Fact]
